Rename test files for log module
